ai_health_advisor_agent/requirements.txt
streamlit==1.32.0
openai==1.70.0
python-dotenv==1.0.1
PyPDF2==3.0.1